Retain big TextPool memory softly
Before the change, the TextPool.DEFAULT_TEXT_POOL would be a strongly
referenced, 300 MB memory block. After the change, it's loaded when
used, and shared opportunistically, by keeping it on a soft reference.
Additionally, a strong reference is kept for 10 seconds to avoid
trashing. On author's laptop, new TextPool gets created in approximately
0.85 s.

This is expected to reduce memory usage for applications that generate
TPC-H data from time to time.